requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
sendgrid>=6.11.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
